requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-yasg==1.21.9
flake8==7.1.2
gunicorn==23.0.0
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.9.0
